Внесены изменения в viewController перед маштабными правками. Новые группы, вынесение структрур из ViewControler
